+++ src/main/java/com/example/diary/domain/schedule/service/ScheduleServiceImpl.java
@@ -1,37 +1,52 @@
 package com.example.diary.domain.schedule.service;
 
 import com.example.diary.domain.member.model.Member;
-import com.example.diary.domain.schedule.controller.request.ScheduleCreateRequestDTO;
-import com.example.diary.domain.schedule.controller.request.ScheduleDeleteRequestDTO;
-import com.example.diary.domain.schedule.controller.request.ScheduleUpdateRequestDTO;
-import com.example.diary.domain.schedule.service.dto.ScheduleCreateDTO;
-import com.example.diary.domain.schedule.service.dto.ScheduleInfoDTO;
-import com.example.diary.domain.schedule.service.dto.ScheduleUpdateDTO;
-import com.example.diary.global.exception.CustomException;
-import com.example.diary.global.exception.ErrorCode;
+import com.example.diary.domain.member.repository.MemberRepository;
+import com.example.diary.domain.schedule.dto.request.ScheduleCreateRequestDTO;
+import com.example.diary.domain.schedule.dto.request.ScheduleDeleteRequestDTO;
+import com.example.diary.domain.schedule.dto.request.ScheduleUpdateRequestDTO;
+import com.example.diary.domain.schedule.dto.service.ScheduleCreateDTO;
+import com.example.diary.domain.schedule.dto.service.ScheduleInfoDTO;
+import com.example.diary.domain.schedule.dto.service.ScheduleUpdateDTO;
 import com.example.diary.domain.schedule.model.Schedule;
 import com.example.diary.domain.schedule.repository.ScheduleRepository;
+import com.example.diary.global.exception.CustomException;
+import com.example.diary.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
 public class ScheduleServiceImpl implements ScheduleService {
-
     private final ScheduleRepository scheduleRepository;
-
+    private final MemberRepository memberRepository;
 
     @Override
     @Transactional
     public void register(ScheduleCreateRequestDTO dto, Member member) {
+        Member assignedMember = null;
+
+        if (dto.assignedMemberId() != null) {
+            assignedMember = memberRepository.findById(dto.assignedMemberId())
+                    .orElseThrow(() ->
+                            new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
+        }
+
+
         ScheduleCreateDTO scheduleCreateDTO = new ScheduleCreateDTO(
                 dto.title(),
                 dto.content(),
                 dto.password(),
-                member
+                false,
+                dto.isPrivate(),
+                member,
+                assignedMember
         );
 
         scheduleRepository.register(scheduleCreateDTO);
@@ -43,6 +58,37 @@ public class ScheduleServiceImpl implements ScheduleService {
                 .stream()
                 .map(ScheduleInfoDTO::from)
                 .toList();
+    }
+
+    @Override
+    public List<ScheduleInfoDTO> searchByTitle(String title) {
+        return scheduleRepository.searchByTitle(title)
+                .stream().map(ScheduleInfoDTO::from)
+                .toList();
+    }
+
+    @Override
+    public Map<Long, List<ScheduleInfoDTO>> findAllByAssignedMember() {
+        Map<Long, List<ScheduleInfoDTO>> map = new HashMap<>();
+        scheduleRepository.findAll()
+                .forEach(schedule -> {
+                    Member assignedMember = schedule.getAssignedMember();
+                    Long id;
+                    if (assignedMember == null) {
+                        id = -1L; // TODO: 어나니머스로 바꿀 예정
+                    } else {
+                        id = assignedMember.getId();
+                    }
+
+                    if (map.containsKey(id)) {
+                        map.get(id).add(ScheduleInfoDTO.from(schedule));
+                    } else {
+                        List<ScheduleInfoDTO> list = new ArrayList<>();
+                        list.add(ScheduleInfoDTO.from(schedule));
+                        map.put(id, list);
+                    }
+                });
+        return map;
     }
 
     @Override
@@ -58,14 +104,29 @@ public class ScheduleServiceImpl implements ScheduleService {
     public ScheduleInfoDTO modifySchedule(ScheduleUpdateRequestDTO dto, Member member) {
         Schedule schedule = scheduleRepository.findById(dto.id())
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
+
+        Member assignedMember = null;
+
+        if (dto.assignedMemberId() != null) {
+            assignedMember = memberRepository.findById(dto.assignedMemberId())
+                    .orElseThrow(() ->
+                            new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
+        }
+
         Schedule updateSchedule = schedule.update(
                 dto.title(),
                 dto.content(),
                 dto.password(),
-                member
+                dto.isDone(),
+                dto.isPrivate(),
+                member,
+                assignedMember
         );
 
-        return ScheduleInfoDTO.from(scheduleRepository.update(dto.id(), ScheduleUpdateDTO.from(updateSchedule)));
+        return ScheduleInfoDTO.from(
+                scheduleRepository.update(dto.id(),
+                        ScheduleUpdateDTO.from(updateSchedule))
+        );
     }
 
     @Override
@@ -74,7 +135,8 @@ public class ScheduleServiceImpl implements ScheduleService {
         Schedule schedule = scheduleRepository.findById(dto.id())
                 .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_EXCEPTION));
 
-        schedule.deleteSchedule(dto.password(), member);
-        scheduleRepository.deleteById(dto.id());
+        if (schedule.hasAuthorization(dto.password(), member)){
+            scheduleRepository.deleteById(dto.id());
+        }
     }
 }
